requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
pytesseract
Pillow
langdetect
scikit-learn
nltk
sqlalchemy
